+++ pkg/codegen/hcl2/model/binder_expression.go
@@ -25,15 +25,34 @@ import (
 	"github.com/zclconf/go-cty/cty/convert"
 )
 
+type BindOption func(options *bindOptions)
+
+func AllowMissingVariables(options *bindOptions) {
+	options.allowMissingVariables = true
+}
+
+type bindOptions struct {
+	allowMissingVariables bool
+}
+
 type expressionBinder struct {
+	options     bindOptions
 	anonSymbols map[*hclsyntax.AnonSymbolExpr]Definition
 	scope       *Scope
 	tokens      _syntax.TokenMap
 }
 
 // BindExpression binds an HCL2 expression using the given scope and token map.
-func BindExpression(syntax hclsyntax.Node, scope *Scope, tokens _syntax.TokenMap) (Expression, hcl.Diagnostics) {
+func BindExpression(syntax hclsyntax.Node, scope *Scope, tokens _syntax.TokenMap,
+	opts ...BindOption) (Expression, hcl.Diagnostics) {
+
+	var options bindOptions
+	for _, opt := range opts {
+		opt(&options)
+	}
+
 	b := &expressionBinder{
+		options:     options,
 		anonSymbols: map[*hclsyntax.AnonSymbolExpr]Definition{},
 		scope:       scope,
 		tokens:      tokens,
@@ -43,12 +62,14 @@ func BindExpression(syntax hclsyntax.Node, scope *Scope, tokens _syntax.TokenMap
 }
 
 // BindExpressionText parses and binds an HCL2 expression using the given scope.
-func BindExpressionText(source string, scope *Scope, initialPos hcl.Pos) (Expression, hcl.Diagnostics) {
+func BindExpressionText(source string, scope *Scope, initialPos hcl.Pos,
+	opts ...BindOption) (Expression, hcl.Diagnostics) {
+
 	syntax, tokens, diagnostics := _syntax.ParseExpression(source, "<anonymous>", initialPos)
 	if diagnostics.HasErrors() {
 		return nil, diagnostics
 	}
-	return BindExpression(syntax, scope, tokens)
+	return BindExpression(syntax, scope, tokens, opts...)
 }
 
 // bindExpression binds a single HCL2 expression.
@@ -155,16 +176,15 @@ func getOperationSignature(op *hclsyntax.Operation) StaticFunctionSignature {
 	for i, p := range ctyParams {
 		sig.Parameters[i] = Parameter{
 			Name: p.Name,
-			Type: InputType(ctyTypeToType(p.Type, p.AllowNull)),
+			Type: ctyTypeToType(p.Type, p.AllowNull),
 		}
 	}
 	if p := op.Impl.VarParam(); p != nil {
 		sig.VarargsParameter = &Parameter{
 			Name: p.Name,
-			Type: InputType(ctyTypeToType(p.Type, p.AllowNull)),
+			Type: ctyTypeToType(p.Type, p.AllowNull),
 		}
 	}
-
 	sig.ReturnType = ctyTypeToType(op.Type, false)
 
 	return sig
@@ -183,8 +203,8 @@ func typecheckArgs(srcRange hcl.Range, signature StaticFunctionSignature, args .
 				diagnostics = append(diagnostics, missingRequiredArgument(param, srcRange))
 			}
 		} else {
-			if !param.Type.ConversionFrom(remainingArgs[0].Type()).Exists() {
-				diagnostics = append(diagnostics, ExprNotConvertible(param.Type, remainingArgs[0]))
+			if !InputType(param.Type).ConversionFrom(remainingArgs[0].Type()).Exists() {
+				diagnostics = append(diagnostics, ExprNotConvertible(InputType(param.Type), remainingArgs[0]))
 			}
 			remainingArgs = remainingArgs[1:]
 		}
@@ -197,8 +217,8 @@ func typecheckArgs(srcRange hcl.Range, signature StaticFunctionSignature, args .
 			diagnostics = append(diagnostics, extraArguments(len(signature.Parameters), len(args), srcRange))
 		} else {
 			for _, arg := range remainingArgs {
-				if !varargs.Type.ConversionFrom(arg.Type()).Exists() {
-					diagnostics = append(diagnostics, ExprNotConvertible(varargs.Type, arg))
+				if !InputType(varargs.Type).ConversionFrom(arg.Type()).Exists() {
+					diagnostics = append(diagnostics, ExprNotConvertible(InputType(varargs.Type), arg))
 				}
 			}
 		}
@@ -225,6 +245,7 @@ func (b *expressionBinder) bindAnonSymbolExpression(syntax *hclsyntax.AnonSymbol
 			Traversal: traversal,
 			SrcRange:  syntax.SrcRange,
 		},
+		Tokens:    _syntax.NewScopeTraversalTokens(traversal),
 		RootName:  "",
 		Parts:     []Traversable{lv},
 		Traversal: traversal,
@@ -284,9 +305,9 @@ func (b *expressionBinder) bindConditionalExpression(syntax *hclsyntax.Condition
 	resultType, _ := UnifyTypes(trueResult.Type(), falseResult.Type())
 
 	// Typecheck the condition expression.
-	signature := StaticFunctionSignature{Parameters: []Parameter{{Name: "condition", Type: InputType(BoolType)}}}
-	typecheckDiags := typecheckArgs(syntax.Range(), signature, condition)
-	diagnostics = append(diagnostics, typecheckDiags...)
+	if InputType(BoolType).ConversionFrom(condition.Type()) == NoConversion {
+		diagnostics = append(diagnostics, ExprNotConvertible(InputType(BoolType), condition))
+	}
 
 	return &ConditionalExpression{
 		Syntax:      syntax,
@@ -438,14 +459,6 @@ func (b *expressionBinder) bindFunctionCallExpression(
 	// Compute the function's signature.
 	signature, sigDiags := function.GetSignature(args)
 	diagnostics = append(diagnostics, sigDiags...)
-
-	// Wrap the function's parameter types in input types s.t. they accept eventual arguments.
-	for i := range signature.Parameters {
-		signature.Parameters[i].Type = InputType(signature.Parameters[i].Type)
-	}
-	if signature.VarargsParameter != nil {
-		signature.VarargsParameter.Type = InputType(signature.VarargsParameter.Type)
-	}
 
 	// Typecheck the function's arguments.
 	typecheckDiags := typecheckArgs(syntax.Range(), signature, args...)
@@ -627,6 +640,18 @@ func (b *expressionBinder) bindRelativeTraversalExpression(
 	if tokens == nil {
 		tokens = _syntax.NewRelativeTraversalTokens(syntax.Traversal)
 	}
+
+	// This occurs with splat expressions.
+	if source, ok := source.(*ScopeTraversalExpression); ok {
+		source.Syntax.Traversal = append(source.Syntax.Traversal, syntax.Traversal...)
+		source.Syntax.SrcRange = hcl.RangeBetween(source.Syntax.SrcRange, syntax.SrcRange)
+		source.Tokens.Traversal = append(source.Tokens.Traversal, tokens.Traversal...)
+
+		source.Traversal = source.Syntax.Traversal
+		source.Parts = append(source.Parts, parts[1:]...)
+		return source, diagnostics
+	}
+
 	return &RelativeTraversalExpression{
 		Syntax:    syntax,
 		Tokens:    tokens,
@@ -654,13 +679,20 @@ func (b *expressionBinder) bindScopeTraversalExpression(
 		for i := range parts {
 			parts[i] = DynamicType
 		}
+
+		var diagnostics hcl.Diagnostics
+		if !b.options.allowMissingVariables {
+			diagnostics = hcl.Diagnostics{
+				undefinedVariable(rootName, syntax.Traversal.SimpleSplit().Abs.SourceRange()),
+			}
+		}
 		return &ScopeTraversalExpression{
 			Syntax:    syntax,
 			Tokens:    tokens,
 			Parts:     parts,
 			RootName:  syntax.Traversal.RootName(),
 			Traversal: syntax.Traversal,
-		}, hcl.Diagnostics{undefinedVariable(rootName, syntax.Traversal.SimpleSplit().Abs.SourceRange())}
+		}, diagnostics
 	}
 
 	parts, diagnostics := b.bindTraversalParts(def, syntax.Traversal.SimpleSplit().Rel)
@@ -686,23 +718,33 @@ func (b *expressionBinder) bindSplatExpression(syntax *hclsyntax.SplatExpr) (Exp
 
 	sourceType := unwrapIterableSourceType(source.Type())
 	elementType := sourceType
-	if arr, isList := sourceType.(*ListType); isList {
-		elementType = arr.ElementType
-	} else if sourceType != DynamicType {
-		source = &TupleConsExpression{
-			Syntax: &hclsyntax.TupleConsExpr{
-				Exprs:     []hclsyntax.Expression{syntax.Source},
-				SrcRange:  syntax.Source.Range(),
-				OpenRange: syntax.Source.StartRange(),
-			},
-			Expressions: []Expression{source},
-			exprType:    NewListType(source.Type()),
+	switch sourceType := sourceType.(type) {
+	case *ListType:
+		elementType = sourceType.ElementType
+	case *SetType:
+		elementType = sourceType.ElementType
+	case *TupleType:
+		elementType, _ = UnifyTypes(sourceType.ElementTypes...)
+	default:
+		if sourceType != DynamicType {
+			source = &TupleConsExpression{
+				Syntax: &hclsyntax.TupleConsExpr{
+					Exprs:     []hclsyntax.Expression{syntax.Source},
+					SrcRange:  syntax.Source.Range(),
+					OpenRange: syntax.Source.StartRange(),
+				},
+				Tokens:      _syntax.NewTupleConsTokens(1),
+				Expressions: []Expression{source},
+				exprType:    NewListType(source.Type()),
+			}
 		}
 	}
 
-	item := &Variable{
-		Name:         "",
-		VariableType: elementType,
+	item := &SplatVariable{
+		Variable: Variable{
+			Name:         "",
+			VariableType: elementType,
+		},
 	}
 	b.anonSymbols[syntax.Item] = item